add repositories

+++ src/main/java/ru/azaytsev/votingrestaurants/user/to/MenuTo.java
@@ -2,11 +2,10 @@ package ru.azaytsev.votingrestaurants.user.to;
 
 import lombok.AllArgsConstructor;
 
-import java.math.BigDecimal;
-
 @AllArgsConstructor
 public class MenuTo {
     String dish;
 
-    BigDecimal price;
+    Double price;
+
 }
